@JeremyRubin 2019-10-29
For those curious, this is a graph of the transactions that are being tested here. The red ones are the "parentless" ones that we create (along with the fake parent I put in to make the graph render nicely.

This graph doesn't show the weights for how many outputs are being consumed by a child from a parent, an inbound edge simply indicates any amount of consumption.


![Txn Graph](https://user-images.githubusercontent.com/886523/67731547-de91ba80-f9b5-11e9-92be-ce306ba5d74f.png)