@JeremyRubin 2019-10-29
Moar data:

I ran #17268 against master across a number of paramters for the number of transactions.

![image](https://user-images.githubusercontent.com/886523/67787134-46cfb300-fa2d-11e9-8bff-ea5dcfa0b1ec.png)

Transactions | Master | Epoch MemPool | Speedup
-- | -- | -- | --
800 | 0.26 | 0.122783 | 2.117556991
2000 | 1.81 | 0.7989 | 2.265615221
5000 | 11.92 | 5.14931 | 2.314873255
8000 | 32.73 | 13.21 | 2.477668433
10000 | 54.1336 | 22.5658 | 2.398922263
12500 | 87.64 | 39.2 | 2.235714286
17500 | 182.268 | 86.34 | 2.11104934
20000 | 248.16 | 118.3 | 2.097717667


What's interesting about the testing setup is if the only modification you make is the number of transactions, it is "monotonic" that is, G(n) is a subgraph of G(n+1) for all n. Adjusting other parameters will not have the same sort of effect, unfortunately, as the construction is not stable.