@JeremyRubin 2019-10-29
Not as useful to look at, but this is the structure with 10,000 nodes and 800. If anyone wants the svg (fully zoomable), can share.

This is made using the twopi shell like layout. The other layout engines choke with this many nodes.

![smaller_hairball](https://user-images.githubusercontent.com/886523/67802255-029ddc00-fa48-11e9-8286-eb339faf7460.png)


![bighairball](https://user-images.githubusercontent.com/886523/67800046-cec0b780-fa43-11e9-8180-4e208d647d1e.png)